+++ R/ts2clm3.R
@@ -64,8 +64,9 @@
 #' the \code{seas} and \code{thresh} outputs will be rounded to. Default is 4. To
 #' prevent rounding set \code{roundClm = FALSE}. This argument may only be given
 #' numeric values or FALSE.
-#' @param returnDF Boolean. The default (\code{FALSE}) returns the results as
-#' type \code{data.table}. Setting it to \code{TRUE} returns a \code{data.frame}.
+#' @param cppDate Boolean (default = \code{FALSE}) to indicate if the pre-compiled
+#' C++ function \code{seqDates} should be used to fill any potential gaps in the given
+#' time series.
 #'
 #' @details
 #' \enumerate{
@@ -158,9 +159,8 @@ ts2clm3 <- function(data,
                     smoothPercentileWidth = 31,
                     clmOnly = FALSE,
                     var = FALSE,
-                    # the version in ts2clm is incorrect
                     roundClm = 4,
-                    returnDF = FALSE) {
+                    cppDate = FALSE) {
   if (missing(climatologyPeriod))
     stop("Oops! Please provide a period (two dates) for calculating the climatology.")
 
@@ -203,13 +203,13 @@ ts2clm3 <- function(data,
   ts_x <- eval(substitute(x), data)
   if (is.null(ts_x) | is.function(ts_x))
     stop(
-      "Please ensure that a column named 't' is present in your data.frame or that you have assigned a column to the 'x' argument."
+      "Please ensure that a column named 't' is present in your data.table or that you have assigned a column to the 'x' argument."
     )
 
   ts_y <- eval(substitute(y), data)
   if (is.null(ts_y) | is.function(ts_y))
     stop(
-      "Please ensure that a column named 'temp' is present in your data.frame or that you have assigned a column to the 'y' argument."
+      "Please ensure that a column named 'temp' is present in your data.table or that you have assigned a column to the 'y' argument."
     )
 
   if (!inherits(ts_x[1], "Date"))
@@ -224,15 +224,28 @@ ts2clm3 <- function(data,
 
   # BEGIN INSERT make_whole_fast >>>
 
+  if(cppDate){
+    date_start <- ts_xy[1, ts_x]
+    date_end <- ts_xy[.N, ts_x]
+    ts_full <- data.table::data.table(ts_x = seqDates(as.integer(format(date_start, "%Y")),
+                                                      as.integer(format(date_start, "%m")),
+                                                      as.integer(format(date_start, "%d")),
+                                                      as.integer(format(date_end, "%Y")),
+                                                      as.integer(format(date_end, "%m")),
+                                                      as.integer(format(date_end, "%d"))))
+    ts_full[, ts_x := as.Date(ts_x)]
+    rm(date_start, date_end)
+  } else  {
+    ts_full <- data.table::data.table(ts_x = seq.Date(ts_xy[1, ts_x],
+                                                      ts_xy[.N, ts_x],
+                                                      "day"))
+  }
+
   feb28 <- 59
 
   .is_leap_year <- function(year) {
     return((year %% 4 == 0 & year %% 100 != 0) | (year %% 400 == 0))
   }
-
-  ts_full <- data.table::data.table(ts_x = seq.Date(ts_xy[1, ts_x],
-                                                    ts_xy[.N, ts_x],
-                                                    "day"))
 
   ts_whole <-
     data.table::merge.data.table(ts_full,
@@ -442,9 +455,7 @@ ts2clm3 <- function(data,
   rm(ts_mat)
 
   if (clmOnly) {
-    if (returnDF) {
-      data.table::setDF(ts_clim)
-    }
+
     return(ts_clim)
 
   } else {
@@ -463,9 +474,7 @@ ts2clm3 <- function(data,
     }
 
     ts_res[, doy := NULL]
-    if (returnDF) {
-      data.table::setDF(ts_res)
-    }
+
     return(ts_res)
   }
 }